test: merge trainer and runtime test files, finish suite lint cleanup
- fold test_trainer_checkpoints/test_early_stopping/test_ema_convention/
  test_resume_lr_override/test_resume_https_checkpoint into test_trainer.py
  and test_runtime_guards/test_runtime_progress_ddp into test_runtime.py,
  all tests moved verbatim (18 and 13 collected IDs unchanged)
- fix the remaining ruff violations across tests/ (import order, unused
  noqa, ClassVar on mutable class defaults, zip strict, raw regex match)
  and reformat the flagged files
- mark tests/integration with pytest.mark.integration so -m selection works
